actualizado falta por terminar

+++ resources/js/index.js
@@ -34,8 +34,8 @@ class Carrito {
                     icon: "error"
                 });
             } else if (producto.stock > 0 && cantidad > 0 && cantidad <= producto.stock) {
-                let item;
-                if (item = this.items.find(item => item.producto.id == producto.id)) {
+                let item = this.items.find(item => item.producto.id == producto.id);
+                if (item) {
                     if (item.cantidad + cantidad > producto.stock) {
                         item.cantidad = producto.stock;
                         Swal.fire({
@@ -52,6 +52,37 @@ class Carrito {
                 actualizarCarrito();
                 Swal.fire({
                     title: "\'" + producto.nombre + "\' x" + cantidad + " agregado al carrito.",
+                    icon: 'success',
+                    position: 'center',
+                    showConfirmButton: false,
+                    timer: 1500
+                });
+            } else {
+                Swal.fire({
+                    title: 'No hay más stock.',
+                    icon: "error"
+                });
+            }
+        });
+    }
+
+    modificarCantidad(idProducto, cantidad) {
+        obtenerProductos((productos) => {
+            let producto = productos.find(producto => producto.id == idProducto);
+            if (!producto) {
+                Swal.fire({
+                    title: 'El producto no existe.',
+                    icon: "error"
+                });
+            } else if (producto.stock > 0 && cantidad > 0 && cantidad <= producto.stock) {
+                let item = this.items.find(item => item.producto.id == producto.id);
+                if (item) {
+                    item.cantidad = cantidad;
+                }
+                sessionStorage.setItem("carrito", JSON.stringify(carrito));
+                actualizarCarrito();
+                Swal.fire({
+                    title: "\'" + producto.nombre + "\' x" + cantidad + " cantidad modificada.",
                     icon: 'success',
                     position: 'center',
                     showConfirmButton: false,
@@ -141,15 +172,85 @@ function mostrarProductos(productos) {
             const agregarButton = document.getElementById(`agregar${producto.id}`);
             const cantidadInput = document.getElementById(`cantidad${producto.id}`);
             agregarButton.addEventListener("click", () => {
-                carrito.agregarItem(producto.id, cantidadInput.value);
+                carrito.agregarItem(producto.id, Number(cantidadInput.value));
             });
         }
     } else {
-        div.innerHTML = "No se encontraron productos.";
+        const resultados = document.getElementById("resultados");
+        resultados.className = "fs-4 fw-bold fst-italic mt-5";
+        // resultados.className = "fs-3 fw-bold fst-italic";
+        // resultados.innerHTML = "No se encontraron resultados.";
     }
 }
 
-function mostrarCarritoTable() {
+// function mostrarCarritoTable() {
+
+//     const tableBody = document.getElementById("tableCarrito");
+//     tableBody.replaceChildren();
+
+//     if (carrito?.items?.length > 0) {
+
+//         for (let item of carrito.items) {
+//             const row = document.createElement("tr");
+//             row.innerHTML = `
+//             <tr>
+//                 <td>
+//                     <img src="${item.producto.imagen}" class="img-fluid w-50" loading="lazy" alt="${item.producto.nombre}">
+//                     <p class="mb-0">${item.producto.nombre}</p>
+//                 </td>
+//                 <td><input id="cantidadCarrito${item.producto.id}" class="text-center" type="number" value="${item.cantidad}" min="1" max="${item.producto.stock}" step="1"></td>
+//                 <td>$${item.producto.precio}</td>
+//                 <td>$${item.producto.precio * item.cantidad}</td>
+//                 <td><button id="quitar${item.producto.id}" type="submit" class="btn btn-danger"><i class="fa-solid fa-trash-can"></i></button></td>
+//             </tr>
+//             `;
+//             tableBody.appendChild(row);
+
+//             // Button para quitar producto del carrito
+//             const quitarButton = document.getElementById(`quitar${item.producto.id}`);
+//             quitarButton.addEventListener("click", () => {
+//                 Swal.fire({
+//                     title: "¿Quitar '" + item.producto.nombre + "'?",
+//                     icon: "question",
+//                     confirmButtonText: "Aceptar",
+//                     showCancelButton: true,
+//                     cancelButtonText: "Cancelar"
+//                 }).then((result) => {
+//                     if (result.isConfirmed) {
+//                         carrito.quitarItem(item.producto.id);
+//                         sessionStorage.setItem("carrito", JSON.stringify(carrito));
+//                         actualizarCarrito();
+//                         // mostrarCarritoTable();
+//                         Swal.fire({
+//                             title: "\'" + item.producto.nombre + '\' ha sido quitado del carrito.',
+//                             icon: "info",
+//                             position: 'center',
+//                             showConfirmButton: false,
+//                             timer: 1500
+//                         })
+//                     }
+//                 })
+//             });
+//             // Input para modificar cantidad 
+//             const cantidadCarritoInput = document.getElementById(`cantidadCarrito${item.producto.id}`);
+//             cantidadCarritoInput.addEventListener("change", () => {
+//                 carrito.agregarItem(item.producto.id, Number(cantidadCarritoInput.value));
+//             });
+//         }
+//         const total = document.getElementById("total");
+//         total.innerHTML = `
+//         <p>TOTAL=$${carrito.calcularTotal()}</p>
+//         <p>TOTAL CON IVA=$${carrito.calcularTotalIva()}</p>
+//          `;
+//     } else {
+//         const total = document.getElementById("total");
+//         total.innerHTML = `
+//         <p>El carrito está vacío.</p>
+//         `;
+//     }
+// }
+
+function mostrarCarritoAlert() {
 
     const tableBody = document.getElementById("tableCarrito");
     tableBody.replaceChildren();
@@ -161,10 +262,13 @@ function mostrarCarritoTable() {
             row.innerHTML = `
             <tr>
                 <td>
-                    <img src="${item.producto.imagen}" class="img-fluid w-50" loading="lazy" alt="${item.producto.nombre}">
+                    <img src="${item.producto.imagen}" class="img-fluid w-25" loading="lazy" alt="${item.producto.nombre}">
                     <p class="mb-0">${item.producto.nombre}</p>
                 </td>
-                <td><input class="text-center" type="number" value="${item.cantidad}" min="1" max="${item.producto.stock}" step="1"></td>
+                <td>
+                    <input id="cantidadCarrito${item.producto.id}" class="text-center" type="number" value="${item.cantidad}" min="1" max="${item.producto.stock}" step="1">
+                    <button id="modificarCantidad${item.producto.id}" type="submit" class="btn btn-info"><i class="fa-solid fa-arrows-rotate"></i></button>
+                </td>
                 <td>$${item.producto.precio}</td>
                 <td>$${item.producto.precio * item.cantidad}</td>
                 <td><button id="quitar${item.producto.id}" type="submit" class="btn btn-danger"><i class="fa-solid fa-trash-can"></i></button></td>
@@ -186,7 +290,7 @@ function mostrarCarritoTable() {
                         carrito.quitarItem(item.producto.id);
                         sessionStorage.setItem("carrito", JSON.stringify(carrito));
                         actualizarCarrito();
-                        mostrarCarrito();
+                        mostrarCarritoAlert();
                         Swal.fire({
                             title: "\'" + item.producto.nombre + '\' ha sido quitado del carrito.',
                             icon: "info",
@@ -196,6 +300,12 @@ function mostrarCarritoTable() {
                         })
                     }
                 })
+            });
+            // Input para modificar cantidad 
+            const cantidadCarritoInput = document.getElementById(`cantidadCarrito${item.producto.id}`);
+            const modificarCantidadButton = document.getElementById(`modificarCantidad${item.producto.id}`);
+            modificarCantidadButton.addEventListener("click", () => {
+                carrito.modificarCantidad(item.producto.id, Number(cantidadCarritoInput.value));
             });
         }
         const total = document.getElementById("total");
@@ -209,133 +319,6 @@ function mostrarCarritoTable() {
         <p>El carrito está vacío.</p>
         `;
     }
-}
-
-function mostrarCarritoResumen() {
-
-    const ul = document.getElementById("carritoResumen");
-    ul.replaceChildren();
-
-    if (carrito?.items?.length > 0) {
-
-        for (let item of carrito.items) {
-            const li = document.createElement("li");
-            li.innerHTML = `
-            <li class="list-group-item d-flex justify-content-between lh-sm">
-                <div>
-                    <h6 class="my-0">${item.producto.nombre}</h6>
-                    <small class="text-muted">${item.cantidad} x $${item.producto.precio}</small>
-                </div>
-                <span class="text-muted">$${item.cantidad * item.producto.precio}</span>
-            </li>
-            `;
-            ul.appendChild(li);
-        }
-        const total = document.createElement("li");
-        total.innerHTML = `
-        <li class="list-group-item d-flex justify-content-between">
-            <span>Total ($)</span>
-            <strong>$${carrito.calcularTotal()}</strong>
-            <span>Total IVA($)</span>
-            <strong>$${carrito.calcularTotalIva()}</strong>
-        </li>
-        `;
-    } else {
-        const total = document.getElementById("total");
-        total.innerHTML = `
-        <p>El carrito está vacío.</p>
-        `;
-    }
-}
-
-function mostrarCarritoAlert() {
-
-    let carritoDiv;
-    if (carrito?.items?.length > 0) {
-        carritoDiv = `
-        <div class="carrito">
-            <div class="container-fluid">
-                <div class="table-responsive">
-                    <table class="table table-light table-striped table-hover align-middle text-center">
-                        <thead class="table-dark text-center">
-                            <tr>
-                                <th scope="col">Artículo</th>
-                                <th scope="col">Cantidad</th>
-                                <th scope="col">Precio</th>
-                                <th scope="col">Subtotal</th>
-                                <th scope="col">Quitar</th>
-                            </tr>
-                        </thead>
-                        <tbody id="carrito">
-            `;
-
-        for (let item of carrito.items) {
-            carritoDiv += `
-            <tr>
-                <td>
-                    <img src="${item.producto.imagen}" class="img-fluid w-50" loading="lazy" alt="${item.producto.nombre}">
-                    <p class="mb-0">${item.producto.nombre}</p>
-                </td>
-                <td><input class="text-center" type="number" value="${item.cantidad}" min="1" max="${item.producto.stock}" step="1"></td>
-                <td>$${item.producto.precio}</td>
-                <td>$${item.producto.precio * item.cantidad}</td>
-                <td><button id="quitar${item.producto.id}" type="submit" class="btn btn-danger"><i class="fa-solid fa-trash-can"></i></button></td>
-            </tr>
-             `;
-            // Button para quitar producto del carrito
-            const quitarButton = document.getElementById(`quitar${item.producto.id}`);
-            quitarButton.addEventListener("click", () => {
-                Swal.fire({
-                    title: "¿Quitar " + item.producto.nombre + "?",
-                    icon: "question",
-                    confirmButtonText: "Aceptar",
-                    showCancelButton: true,
-                    cancelButtonText: "Cancelar"
-                }).then((result) => {
-                    if (result.isConfirmed) {
-                        carrito.quitarItem(item.producto.id);
-                        sessionStorage.setItem("carrito", JSON.stringify(carrito));
-                        actualizarCarrito();
-                        Swal.fire({
-                            title: "\'" + item.producto.nombre + '\' ha sido quitado del carrito.',
-                            icon: "info",
-                            position: 'center',
-                            showConfirmButton: false,
-                            timer: 1500
-                        })
-                    }
-                })
-            });
-        }
-
-        carritoDiv += `
-                </tbody>
-            </table>
-        </div>
-        `;
-        carritoDiv += `
-                <p>TOTAL=$${carrito.calcularTotal()}</p>
-                <p>TOTAL CON IVA=$${carrito.calcularTotalIva()}</p>
-            </div>
-            <div>
-                <a href="./pages/carrito.html">
-                    PAGAR
-                </a>
-            </div>
-        </div>
-        `;
-    } else {
-        carritoDiv = `
-                <p>El carrito está vacío.</p>
-            </div>
-        </div>
-        `;
-    }
-
-    Swal.fire({
-        title: "Carrito",
-        html: carritoDiv
-    });
 }
 
 function actualizarCarrito() {
@@ -372,29 +355,6 @@ function buscarProductos(nombreProducto) {
     });
 }
 
-function generarPago() {
-    Swal.fire({
-        title: "¿Realizar pago por $" + carrito.calcularTotalIva() + "?",
-        icon: "question",
-        confirmButtonText: "Aceptar",
-        showCancelButton: true,
-        cancelButtonText: "Cancelar"
-    }).then((result) => {
-        if (result.isConfirmed) {
-            carrito.vaciarCarrito();
-            sessionStorage.clear();
-            actualizarCarrito();
-            Swal.fire({
-                title: "Pago realizado.",
-                icon: "success",
-                position: 'center',
-                showConfirmButton: false,
-                timer: 1500
-            })
-            // redirigir
-        }
-    })
-}
 
 const buscadorInput = document.getElementById("buscadorInput");
 const buscadorFormulario = document.getElementById("buscadorFormulario");
@@ -409,34 +369,68 @@ carritoButton.onclick = () => {
     mostrarCarritoAlert();
 };
 
-const verCarritoButton = document.getElementById("verButton");
-verCarritoButton.onclick = () => {
-    actualizarCarrito();
-    mostrarCarritoTable();
-};
+// const verCarritoButton = document.getElementById("verButton");
+// verCarritoButton.onclick = () => {
+//     actualizarCarrito();
+//     mostrarCarritoTable();
+// };
 
 const vaciarButton = document.getElementById("vaciarButton");
 vaciarButton.onclick = () => {
-    carrito.vaciarCarrito();
-    sessionStorage.clear();
-    mostrarCarritoTable();
+    if (carrito?.items?.length > 0) {
+        Swal.fire({
+            title: "¿Vaciar carrito?",
+            icon: "question",
+            confirmButtonText: "Aceptar",
+            showCancelButton: true,
+            cancelButtonText: "Cancelar"
+        }).then((result) => {
+            if (result.isConfirmed) {
+                carrito.vaciarCarrito();
+                sessionStorage.clear();
+                mostrarCarritoAlert();
+                actualizarCarrito();
+                Swal.fire({
+                    title: "Carrito vaciado.",
+                    icon: "info",
+                    position: 'center',
+                    showConfirmButton: false,
+                    timer: 1500
+                })
+            }
+        })
+    }
 };
 
-const pagoFormulario = document.getElementById("pagoFormulario");
-pagoFormulario.addEventListener("submit", (e) => {
-    e.preventDefault();
-    generarPago();
-});
-
-
+obtenerProductos(mostrarProductos);
 // Carrito en SessionStorage
 const carrito = new Carrito();
 actualizarCarrito();
 
-obtenerProductos(mostrarProductos);
 
+// Get the modal
+const modal = document.getElementById("carritoModal");
 
+// Get the button that opens the modal
+const btn = document.getElementById("carritoButton");
 
-// Carrito.html
-mostrarCarritoTable();
-mostrarCarritoResumen();
+// Get the <span> element that closes the modal
+const span = document.getElementsByClassName("close")[0];
+
+// When the user clicks on the button, open the modal
+btn.onclick = function () {
+    modal.style.display = "block";
+    mostrarCarritoAlert();
+}
+
+// When the user clicks on <span> (x), close the modal
+span.onclick = function () {
+    modal.style.display = "none";
+}
+
+// When the user clicks anywhere outside of the modal, close it
+window.onclick = function (event) {
+    if (event.target == modal) {
+        modal.style.display = "none";
+    }
+}
